+++ src/stores/Frame/FrameStore.ts
@@ -198,7 +198,6 @@ export class FrameStore {
 
     @observable stokesFiles: CARTA.StokesFile[];
 
-    @observable isPreview: boolean;
     @observable previewViewWidth: number;
     @observable previewViewHeight: number;
     @observable previewPVRasterData: Float32Array;
@@ -1152,6 +1151,10 @@ export class FrameStore {
         return getFormattedWCSPoint(this.wcsInfoForTransformation, this.center);
     }
 
+    @computed get isPreview(): boolean {
+        return this.frameInfo.preview;
+    }
+
     @computed get previewCursorValue(): {position: Point2D; channel: number; value: number} | null {
         if (!this.isPreview) {
             return null;
@@ -1223,9 +1226,7 @@ export class FrameStore {
 
         this.stokesFiles = [];
 
-        this.isPreview = frameInfo.preview;
-
-        this.distanceMeasuring = !frameInfo.preview && new DistanceMeasuringStore();
+        this.distanceMeasuring = frameInfo.preview ? null : new DistanceMeasuringStore();
 
         this.dirAxis = -1;
         this.dirAxisSize = -1;